+++ Tests/Unit/Module/Convert.Main.tests.ps1
@@ -194,6 +194,14 @@ try
                         'TechnologyRole' = 'PowerPoint2013'
                     }
                 )
+                'Excel2013' = @(
+                    @{
+                        'id' = 'Windows_All_Excel2013'
+                        'Technology' = 'Windows'
+                        'TechnologyVersion' = 'All'
+                        'TechnologyRole' = 'Excel2013'
+                    }
+                )
             }
             foreach ($sampleString in $sampleStrings.GetEnumerator())
             {